Improvements after review. Related to Issue [WTI-818]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@5237 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 0930132f680e52380d728903356ac6a0f7fa3e4c
Former-commit-id: ec5de7c7499a23d131960a7770bab67bb04d389c

+++ Ringtoets/Revetment/src/Ringtoets.Revetment.Service/IWaveConditionsCalculationService.cs
@@ -57,7 +57,7 @@ namespace Ringtoets.Revetment.Service
                                        double a,
                                        double b,
                                        double c,
-                                       double norm,
+                                       int norm,
                                        WaveConditionsInput input,
                                        string hlcdDirectory,
                                        string ringId,
